@@ -56,7 +56,7 @@
   "unpkg": "dist/cytoscape.min.js",
   "jsdelivr": "dist/cytoscape.min.js",
   "scripts": {
-    "lint": "eslint src benchmark",
+    "lint": "eslint src/**/*.mjs",
     "build": "rollup -c",
     "build:esm": "cross-env FILE=esm rollup -c",
     "build:esm.min": "cross-env FILE=esm.min rollup -c",
@@ -97,6 +97,7 @@
   "devDependencies": {
     "@babel/core": "^7.26.0",
     "@babel/preset-env": "^7.26.0",
+    "@eslint/js": "^9.18.0",
     "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-commonjs": "^28.0.2",
     "@rollup/plugin-node-resolve": "^16.0.0",
@@ -104,13 +105,14 @@
     "@rollup/plugin-terser": "^0.4.4",
     "benchmark": "^2.1.4",
     "bluebird": "^3.5.0",
-    "chai": "^4.1.2",
+    "chai": "^5.1.2",
     "cpy-cli": "^5.0.0",
     "cross-env": "^7.0.0",
     "download-cli": "^1.0.5",
-    "eslint": "^6.0.0",
+    "eslint": "^9.18.0",
     "gh-pages": "^5.0.0",
     "gl-matrix": "^3.4.3",
+    "globals": "^15.14.0",
     "handlebars": "^4.7.6",
     "heap": "^0.2.7",
     "highlight.js": "^10.0.0",
