@@ -1,5 +1,6 @@
-var expect = require('chai').expect;
-var cytoscape = require('../src/test.js', cytoscape);
+import { expect } from 'chai';
+import cytoscape from '../src/test.mjs';
+
 var $$ = cytoscape;
 
 var is = {
