@@ -1,6 +1,8 @@
-var expect = require('chai').expect;
-var cytoscape = require('../src/test.js', cytoscape);
-var Promise = Promise || require('bluebird');
+import { expect } from 'chai';
+import cytoscape from '../src/test.mjs';
+import Bluebird from 'bluebird';
+
+var Promise = Promise || Bluebird;
 
 describe('Core initialisation', function(){
 
