@@ -1,8 +1,8 @@
 import * as util from '../util/index.mjs';
 import * as is from '../is.mjs';
-import get from 'lodash/get';
-import set from 'lodash/set';
-import toPath from 'lodash/toPath';
+import get from 'lodash/get.js';
+import set from 'lodash/set.js';
+import toPath from 'lodash/toPath.js';
 
 let define = {
 
