@@ -1,4 +1,4 @@
-import { expandBoundingBoxSides, makeBoundingBox } from '../../src/math'
+import { expandBoundingBoxSides, makeBoundingBox } from '../../src/math.mjs'
 import { describe } from 'mocha'
 import { expect } from 'chai'
 
